docs: listpays timestamps are u32
The xpay listpays helper emits created_at with json_add_u32 (from
u32 pm->timestamp):

  plugins/xpay/listpays.c:168
    json_add_u32(ret, "created_at", pm->timestamp);

The legacy listpays path serializes both timestamps through
json_add_payment_fields(), which uses json_add_u32 from the u32
wallet_payment struct members:

  lightningd/pay.c:158,160
    json_add_u32(response, "created_at", t->timestamp);
    json_add_u32(response, "completed_at", *t->completed_at);
  wallet/wallet.h:363-364
    u32 timestamp;
    u32 *completed_at;

Changelog-None

+++ cln-rpc/src/model.rs
@@ -9530,7 +9530,7 @@ pub mod responses {
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub bolt12: Option<String>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
-	    pub completed_at: Option<u64>,
+	    pub completed_at: Option<u32>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
 	    pub created_index: Option<u64>,
 	    #[serde(skip_serializing_if = "Option::is_none")]
@@ -9550,7 +9550,7 @@ pub mod responses {
 	    // Path `ListPays.pays[].status`
 	    pub status: ListpaysPaysStatus,
 	    pub amount_sent_msat: Amount,
-	    pub created_at: u64,
+	    pub created_at: u32,
 	    pub payment_hash: Sha256,
 	}
